requirements.txt
requests
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
fastapi
uvicorn[standard]
pydantic